chore: refactor a bit more
